+++ server/__tests__/integration/utils/wasm-loader.ts
@@ -84,6 +84,10 @@ export const WASM_TEST_BINARIES = {
     },
     redirect: {
       redirect: 'redirect.wasm',
+      redirectExtraHeaders: 'redirect-extra-headers.wasm',
+    },
+    emptyEnv: {
+      emptyEnv: 'empty-env.wasm',
     },
     variablesAndSecrets: {
       variablesAndSecrets: 'variables-and-secrets.wasm',